Move LegacySourceManager to legacy package
Separate it from wallpaper specific code
