+++ main/src/main/java/com/google/android/apps/muzei/tasker/TaskerActionReceiver.kt
@@ -20,9 +20,9 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import androidx.core.os.bundleOf
+import com.google.android.apps.muzei.legacy.LegacySourceManager
 import com.google.android.apps.muzei.sync.ProviderManager
 import com.google.android.apps.muzei.util.goAsync
-import com.google.android.apps.muzei.wallpaper.LegacySourceManager
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.twofortyfouram.locale.api.Intent.ACTION_FIRE_SETTING
 import com.twofortyfouram.locale.api.Intent.EXTRA_BUNDLE
