+++ main/src/main/java/com/google/android/apps/muzei/widget/MuzeiAppWidgetProvider.kt
@@ -22,8 +22,8 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.core.os.bundleOf
+import com.google.android.apps.muzei.legacy.LegacySourceManager
 import com.google.android.apps.muzei.util.goAsync
-import com.google.android.apps.muzei.wallpaper.LegacySourceManager
 import com.google.firebase.analytics.FirebaseAnalytics
 
 /**
